refactor: decentralize component types
